Fix processing of timeouts when duplicate request exists
When processing a timed out request it is possible that a new
request with the same ID has already been made.  This could be
because the original request has already been processed by
we are only just seeing the timer go.  Or because IDs are not
sufficiently unique (for examle with SMS pings whose ID is the
recipient

To handle this we added an 'isProcessed' flag to requests.  If a request
is marked is isProcessed then we just toss it out.  If instead
we end up with a request whose timeout comes in but is not longer listed
then we mark the request as 'errored' by callig processError